+++ vite.config.ts
@@ -24,6 +24,8 @@ export default defineConfig({
             { find: "tardigrade-store/persist", replacement: path.resolve(__dirname, 'sources/persist/index.ts') },
             { find: "tardigrade-store/history", replacement: path.resolve(__dirname, 'sources/history/index.ts') },
             { find: "tardigrade-store/ward", replacement: path.resolve(__dirname, 'sources/ward/index.ts') },
+            { find: "tardigrade-store/vue", replacement: path.resolve(__dirname, 'sources/vue/index.ts') },
+            { find: "tardigrade-store/svelte", replacement: path.resolve(__dirname, 'sources/svelte/index.ts') },
             { find: "tardigrade-store", replacement: path.resolve(__dirname, 'sources/index.ts') },
         ],
     },